Make button subscription requests just once
There was an issue related to sending of the same button
subscription requests every time application changes HMI
state. Even if application was able to subscribe to these buttons
successfully first time, it will definitely receive IGNORED error
from SDL on all subsequent attempt to do that.
A new flag has been added to make sending of such requests
just once like for AddCommand.
Additionally, `.catch()` has been added to gracefully handle the
errorneous response from SDL and perform the rest of code.
Also, RC buttons was replaced with the regular buttons as this
is not supposed to be an RC application.